camera/requirements.txt
pyzmq
structlog
